feat(backend): extract dependency graph logic and integrate into deployment pipeline (#495)
